requirements.txt
streamlit
google-generativeai
feedparser
PyGithub
pandas
